Exchange por fazer: resolver problema no envio de listas e na receção da classe Oferta

+++ exchange/src/main/java/Leilao.java
@@ -3,18 +3,18 @@ import java.util.Map;
 
 public class Leilao {
     private String empresa;
-    private float montante;
-    private float taxaMaxima;
+    private double montante;
+    private double taxaMaxima;
     private Map<String, Oferta> investidores;
 
-    public Leilao(String empresa, float montante, float taxaMaxima){
+    public Leilao(String empresa, double montante, double taxaMaxima){
         this.empresa = empresa;
         this.montante = montante;
         this.taxaMaxima = taxaMaxima;
         this.investidores = new HashMap<>();
     }
 
-    public Leilao(String empresa, float montante, float taxaMaxima, Map<String, Oferta> inv){
+    public Leilao(String empresa, double montante, double taxaMaxima, Map<String, Oferta> inv){
         this.empresa = empresa;
         this.montante = montante;
         this.taxaMaxima = taxaMaxima;
@@ -25,11 +25,11 @@ public class Leilao {
         return empresa;
     }
 
-    public float getMontante() {
+    public double getMontante() {
         return montante;
     }
 
-    public float getTaxaMaxima() {
+    public double getTaxaMaxima() {
         return taxaMaxima;
     }
