+++ service/src/main/java/diretorio/representations/OfertaRepresentation.java
@@ -4,11 +4,11 @@ import com.fasterxml.jackson.annotation.JsonCreator;
 import com.fasterxml.jackson.annotation.JsonProperty;
 
 public class OfertaRepresentation {
-    public final float montante;
-    public final float taxa;
+    public final double montante;
+    public final double taxa;
 
     @JsonCreator
-    public OfertaRepresentation(@JsonProperty("montante") float m, @JsonProperty("taxa") float t) {
+    public OfertaRepresentation(@JsonProperty("montante") double m, @JsonProperty("taxa") double t) {
         this.montante = m;
         this.taxa = t;
     }
